Refactor Node options
Node options are now done by traits.

The reasoning behind this is that one wants to inherit options from
nodes they subclass and that lookup of these options should be fast.

Lookup of views is this:
  [Innate.options.roots, Innate.options.views,
    Node.trait[:views], name]

Lookup of layouts is this:
  [Innate.options.roots, Innate.options.layouts,
    Node.trait[:layouts], name]

If the traits are empty arrays, the Node.mapping is used instead.

Setting a custom view or layout mapping works like:

  class Foo
    Innate.node('/')
    map_views '/foobar'
    map_layouts '/foobar'
  end

More about this soon in the docs.